+++ qgreenland/config/layers/Places/populated_places.py
@@ -44,6 +44,11 @@ QGreenland Team - Noted Data Issues:
             dataset=statbank,
             asset=statbank.assets["international_passengers"],
         ),
+        # This input provides numbers of cruise passengers (csv)
+        LayerInput(
+            dataset=statbank,
+            asset=statbank.assets["cruise_passengers"],
+        ),
         # This input provides a multipolygon of municipalities and gives us the
         # municipality name for each populated place. (gpkg)
         LayerInput(