+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # v4.0.0alpha4 (TBD)
 
 - Added layers:
-  - "Human activity/Airport international passengers (temporal)"
+  - "Human activity/"
+    - "Airport international passengers (temporal)"
+    - "Cruise passengers by harbor (temporal)"
 
 # v4.0.0alpha3 (2025-03-26)